Refactored the code to avoid code duplication
